refactor(avatar): extract shared styling into module
UserAvatarComponent and Viral::AvatarComponent duplicated size
mappings, font styles, and HSLA colour generation. Share them via
AvatarStyling so both stay in sync.
