+++ app/components/user_avatar_component.rb
@@ -2,15 +2,9 @@
 
 # Displays a user's initials on a deterministic coloured background.
 class UserAvatarComponent < Component
-  attr_reader :user, :initials, :size, :font_styles
+  include AvatarStyling
 
-  SIZE_DEFAULT = :medium
-  SIZE_MAPPINGS = {
-    xs: 'w-6 h-6 rounded-sm',
-    small: 'w-8 h-8',
-    medium: 'w-12 h-12',
-    large: 'w-16 h-16'
-  }.freeze
+  attr_reader :user, :initials, :size, :font_styles
 
   def initialize(user:, size: SIZE_DEFAULT, decorative: false, **system_arguments)
     @user = user
@@ -37,23 +31,5 @@ class UserAvatarComponent < Component
         opts[:class]
       )
     end
-  end
-
-  private
-
-  def font_styles_for(size)
-    if %i[xs small].include?(size)
-      'text-sm'
-    else
-      size == :medium ? 'text-2xl font-semibold' : 'text-4xl font-bold'
-    end
-  end
-
-  def generate_hsla_colour(seed)
-    h = Digest::MD5.hexdigest(seed).to_i(16) % 360
-    {
-      border: "hsla(#{h}, 100%, var(--tw-avatar-border-lightness), var(--tw-avatar-bg-alpha))",
-      background: "hsla(#{h}, 100%, var(--tw-avatar-bg-lightness), var(--tw-avatar-bg-alpha))"
-    }
   end
 end
